Show doors and tileswaps on maps. Fix warnings

+++ src/Chests.cpp
@@ -6,7 +6,7 @@ Chests::Chests(const std::vector<uint8_t>& offsets, const std::vector<uint8_t>& 
 	int max_offset = offsets[0];
 	int last_room = 0;
 	std::map<uint16_t, uint8_t> chests;
-	for (int i = 0; i < offsets.size() - 1; ++i)
+	for (int i = 0; i < static_cast<int>(offsets.size() - 1); ++i)
 	{
 		if (offsets[i] == 0 && max_offset > 0)
 		{
@@ -23,9 +23,10 @@ Chests::Chests(const std::vector<uint8_t>& offsets, const std::vector<uint8_t>& 
 			last_room = i;
 		}
 	}
-	if (max_offset < contents.size())
+	if (max_offset < static_cast<int>(contents.size()))
 	{
-		chests.insert({ offsets.size() - 1, contents.size() - max_offset });
+		chests.insert({ static_cast<uint16_t>(offsets.size() - 1),
+			            static_cast<uint8_t>(contents.size() - max_offset) });
 	}
 	int cur_offset = 0;
 	for (const auto& c : chests)
@@ -132,7 +133,7 @@ void Chests::ClearRoomNoChestsFlag(uint16_t room)
 bool Chests::CleanupRoomChests(const GameData& gd)
 {
 	int chest_count = 0;
-	for (int r = 0; r < gd.GetRoomData()->GetRoomCount(); ++r)
+	for (uint16_t r = 0; r < static_cast<uint16_t>(gd.GetRoomData()->GetRoomCount()); ++r)
 	{
 		auto ents = gd.GetSpriteData()->GetRoomEntities(r);
 		int chests = 0;
